% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logticks.R
\name{add_logticks}
\alias{add_logticks}
\title{Add logticks}
\usage{
add_logticks(
  base = 10,
  sides = "bl",
  scaled = TRUE,
  short = unit(0.1, "cm"),
  mid = unit(0.2, "cm"),
  long = unit(0.3, "cm"),
  colour = "black",
  size = 0.5,
  linetype = 1,
  alpha = 1,
  color = NULL,
  data = data.frame(x = NA),
  ...
)
}
\arguments{
\item{base}{Base}

\item{sides}{Sides}

\item{scaled}{Scaled}

\item{short}{Short}

\item{mid}{Mid}

\item{long}{Long}

\item{colour}{Colour}

\item{size}{Size}

\item{linetype}{Linetype}

\item{alpha}{Alpha}

\item{color}{Color}

\item{data}{Data}

\item{...}{Other}
}
\value{
Added logticks
}
\description{
Add logticks
}
