% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/around_point.R
\name{around_point}
\alias{around_point}
\title{Subset around point}
\usage{
around_point(data, var, ref_condition, time_radius)
}
\arguments{
\item{data}{Data frame}

\item{var}{Variable to center around}

\item{ref_condition}{Reference point (e.g. "Day"/"Night")}

\item{time_radius}{Time around point}
}
\value{

}
\description{
Subset around point
}
