% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_video.R
\name{assign_video}
\alias{assign_video}
\title{Assign video}
\usage{
assign_video(track_list, animal_ids)
}
\arguments{
\item{track_list}{List from video files}

\item{animal_ids}{Animal ids}
}
\value{
track_list
}
\description{
Assign video
}
