% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_states.R
\name{classify_states}
\alias{classify_states}
\title{Classify states Across Time-scales}
\usage{
classify_states(data, movement_var, window_width, .keep = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{window_width}{Window width in number of frames}

\item{group}{}
}
\description{
Classify states Across Time-scales
}
