% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_observations.R
\name{compress_observations}
\alias{compress_observations}
\title{Compress observations}
\arguments{
\item{data}{Data}

\item{n_observations}{Number of observations in output data frame}

\item{group}{grouping variable}
}
\value{
Smaller data frame where values are means of N observations
}
\description{
Get an average of your data where the output is n rows (approximately).
}
