% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_classifications.R
\name{filter_classifications}
\alias{filter_classifications}
\title{Filter Classifications}
\usage{
filter_classifications(data, window_size, group)
}
\arguments{
\item{group}{}
}
\description{
Filter Classifications
}
