% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_forward_backward.R
\name{filter_forward_backward}
\alias{filter_forward_backward}
\title{Forward-backward Filter}
\usage{
filter_forward_backward(data, movement_var, window_width)
}
\arguments{
\item{data}{Data frame}

\item{group}{}
}
\description{
Forward-backward Filter
}
