% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_state_numbers.R
\name{generete_state_numbers}
\alias{generete_state_numbers}
\title{Generate bout numbers for binary variable}
\usage{
generete_state_numbers(data, var, .keep = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{var}{Variable name (in quotes)}

\item{.keep}{Keep both intermediate filter components (forward/backward)}
}
\value{
Data frame with classifications
}
\description{
Generate bout numbers for binary variable
}
