% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_periods.R
\name{identify_periods}
\alias{identify_periods}
\title{IDENTIFY PERIODS OF INTEREST}
\usage{
identify_periods(data, var, newvar, threshold = "minmax", as_factor = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{var}{Variable to be identified}

\item{newvar}{New variable name}

\item{threshold}{Method for thresholding. Either "minmax", "mediansd" or a number.}

\item{as_factor}{Whether an output should be returned as factor}
}
\value{

}
\description{
IDENTIFY PERIODS OF INTEREST
}
