% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_ecdf.R
\name{inverse_ecdf}
\alias{inverse_ecdf}
\title{inverse_ecdf}
\usage{
inverse_ecdf(data, col, min_length, rest_state = FALSE)
}
\arguments{
\item{data}{A data frame (e.g. a tibble)}

\item{col}{Variable to make the density function for.}

\item{min_length}{Minimal starting duration in frames}

\item{rest_state}{Whether to make the ecdf for resting or active periods. TRUE or FALSE (default)}
}
\description{
inverse_ecdf
}
