% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_moving.R
\name{is_moving}
\alias{is_moving}
\title{is_moving}
\usage{
is_moving(df, type, n_frames = 50)
}
\arguments{
\item{df}{Data frame.}

\item{type}{Use mean or median.}

\item{n_frames}{Number of frames used to generate rolling average.}
}
\value{

}
\description{
is_moving
}
