% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_time.R
\name{join_timeseries}
\alias{join_timeseries}
\title{Ensure that grouping levels match across data frames - add an error/warning if not.}
\usage{
join_timeseries(
  .x,
  y,
  by = NULL,
  group = NULL,
  method = NA,
  copy = FALSE,
  suffix = c("", ".y"),
  keep = NULL
)
}
\description{
Ensure that grouping levels match across data frames - add an error/warning if not.
}
