% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_videos.R
\name{merge_videos}
\alias{merge_videos}
\title{Merge videos}
\usage{
merge_videos(df_list)
}
\arguments{
\item{df_list}{List of data frames}
}
\description{
Merges data sets for consecutive videos for each animal.
}
