% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_align_all_data.R
\name{read_align_all_data}
\alias{read_align_all_data}
\title{Title}
\usage{
read_align_all_data(filenames, bin_seconds)
}
\arguments{
\item{filenames}{Filenames}

\item{bin_seconds}{Bin duration in seconds}
}
\value{

}
\description{
Title
}
