% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_align_data.R
\name{read_align_data}
\alias{read_align_data}
\title{Title}
\usage{
read_align_data(data, bin_seconds, time, animal_id)
}
\arguments{
\item{bin_seconds}{Bin duration in seconds}

\item{filename}{Filenames}
}
\value{
Aligned data frame
}
\description{
Title
}
