% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_switch_data.R
\name{read_switch_data}
\alias{read_switch_data}
\title{Title}
\usage{
read_switch_data(filenames, bin_seconds, time_radius = FALSE)
}
\arguments{
\item{filenames}{Filenames}

\item{bin_seconds}{Bin duration in seconds}

\item{time_radius}{Radius around change}
}
\value{

}
\description{
Title
}
