% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/around_point.R
\name{around_point}
\alias{around_point}
\title{Subset around point}
\usage{
around_point(data, var, from, time_var, time_radius = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{var}{Variable to center around}

\item{from}{Reference point (e.g. "Day"/"Night")}

\item{time_var}{Which variable represents time}

\item{time_radius}{Time around point}
}
\value{
Subset of data frame around point in time
}
\description{
Subset around point
}
