% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_states.R
\name{classify_states}
\alias{classify_states}
\title{Classify states Across Time-scales}
\usage{
classify_states(data, movement_var, window_widths, .keep = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{movement_var}{Binary (1/0) variable to be used for the classification}

\item{window_widths}{Window width for the filter}

\item{.keep}{Keep both intermediate filter components (forward/backward)}
}
\value{
Data frame with classifications
}
\description{
Classify states Across Time-scales
}
