% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_observations.R
\name{compress_observations}
\alias{compress_observations}
\title{Compress observations}
\arguments{
\item{data}{Data}

\item{vars}{Variables to mean}

\item{n_observations}{Number of observations in output data frame}

\item{group}{grouping variable}
}
\value{

}
\description{
Get an average of your data where the output is n rows (approximately).
}
