% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_classifications.R
\name{filter_classifications}
\alias{filter_classifications}
\title{Filter Classifications}
\usage{
filter_classifications(data, window_widths)
}
\arguments{
\item{data}{Data frame}

\item{window_widths}{Minimum duration of bout}
}
\value{
Filtered data frame
}
\description{
Filter Classifications
}
