% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_position.R
\name{find_position}
\alias{find_position}
\title{Find position}
\usage{
find_position(df, exp_setup = c("wellplate", "tube"), animal_ids)
}
\arguments{
\item{df}{track_list of data frames.}

\item{exp_setup}{Experimental setup, eith "wellplate " or "tube".}

\item{animal_ids}{Vector of animal IDs, as strings.}
}
\value{
A single tibble
}
\description{
Find position
}
