% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_align_data.R
\name{read_align_data}
\alias{read_align_data}
\title{Title}
\usage{
read_align_data(filename, bin_seconds)
}
\arguments{
\item{filename}{Filenames}

\item{bin_seconds}{Bin duration in seconds}
}
\value{
Aligned data frame
}
\description{
Title
}
