% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_time.R
\name{sync_time}
\alias{sync_time}
\title{Sync time}
\usage{
sync_time(.data, x, y, data2)
}
\arguments{
\item{.data}{Data frame or tibble}

\item{x}{Parameter used to merge (often time)}

\item{y}{Variable to be synced along with x}

\item{data2}{The second data frame}
}
\value{
Tibble
}
\description{
Sync time
}
